Syracusae VIII (CIL X 7130)
Notes:
These fragmentary inscriptions are related to the loca of the amphitheatre of Syracuse and are part of CIL 10.7130. Three fragments (7130.1,23,24) are not among those presented here (maybe in ISic they are not consecutive to these?  ISic0753 does not exist, so it could be used to add at least 7130.23 if it is not yet in ISicily. Let's see when I will finish the inscriptions from Syracuse).

Encoding text division, cleaning up the autogenerated text and checking the edition.

Points to check:
- see if you are happy with the bibliography formatting (tav., mention, etc.)
- Manganaro 1989 proposes to join
(a) the fragments CIL X 7130.1+2 (ISic0732+another fragm. maybe not encoded yet?);
(b) the fragments CIL X 7130.5+4 (ISic0735+0734);
(c) the fragments CIL X 7130.13+15 (ISic0743+0745). I reported the information in the commentary but I used Buonocore's edition which reports the fragments separately.