Updated
there is still something wrong with the subdivision. In the case of impressions, if we want to use the layout of the main body to describe the impression, we must have measurement options there that distinguish the epigraphic field (impression) from the object, in this case the diameter (We can perhaps add it in the handnotes together with height of the letters?) and we must also have the condition option because it can be different from the object.