Update ISic003710
We have here the case of a text that could be divided into two parts/uses (depending on how we interpret it): the first two lines could indeed be a funerary text, while the next could be a writing exercise (probably) made when the document was no longer serving its purpose, thus as reuse.
In addition to the comment, I would propose to include the information on the writing exercise as an additional category of the type of inscription. However, I cannot find the ref in eagle...