Seven inscriptions
Update of two from last week, and five new ones. I opted to stay faithful to CIL in transcription, and explain Manganaro's and Eck's conjectures in apparatus. I hope that's all OK. Slightly wobbly on on ISic374, where there appears to be a deliberate hole in the middle of the text. I have coded it as I interpreted the Italian text in CIL, but I may have got the wrong end of the stick!