Updating 3 inscriptions
ISic000695, 000696, 000760. There were some issues with capturing the expansion of Aemiliam and Dalmatiam in 000696, since half of the word is lost and so the <supplied> interferes with the <expan> elements. I was unable to put the full <expan> brackets around Aemiliam, and so only did it for the miliam, although I'm not sure if that's standard procedure.