Update Camarina>Santa Croce Camerina?
I think we should change all references to 'Camarina' to 'Santa Croce Camerina', which is actually the name of the modern municipality in which the ancient site is located (and therefore insert the coordinates for the generic modern site as http://sws.geonames.org/654085). If you agree, I can proceed with all the other files and change the reference in the places list too.