preliminary edits
I've made some preliminary edits to 537 (and added images), and just a couple of very minor cleaning edits to 3364 and 3644, and resolved a conflict on the taxonomy file. 3364 and 3644 are fine to be merged, and the taxonomy file; Simona and I plan to discuss and tweak 537 a little further. No action required on your part!