Edits on Elymian inscriptions after autopsy (VII)
About the white painting, we could also add somethng like: "The letters have been highlighted with white painting to make them more legible in modern times, probably in the 1970s when the pieces were discovered. No trace of the operation remains in the archives. That the painting is a modern intervention can be inferred from the presence of it even in the broken parts of the vase (continuing the line on the actual engraving). [I can try sth clearer here]. Only the Elymian documents in which the inscription was clearly legible were spared." [suggestions and rewording are welcome!].