Update ISic000655.xml
various minor edits / corrections (missing 'p' on layout description; current location is described at end of provenance (we only do one provenance element, besides autopsy, at present); added `<w part="I">` to the two letters in line 5 - letters that can't be resolved should either be tagged with orig if completely unresolveable, or with a word part if known to be beginning or end of a word. (I=initial, F=final).